refactor(MP-71): Add some fixes after code review.
