+++ src/components/molecules/Table/Content/index.tsx
@@ -1,6 +1,5 @@
 import { TableInstance } from 'react-table';
 import { Box, Table } from '@chakra-ui/react';
-import React from 'react';
 import { BaseObject } from 'model/base-object';
 import Header from 'components/molecules/Table/Header';
 import Body from 'components/molecules/Table/Body';
